pyproject.toml
[tool.poetry]
authors = ["Artsiom Kvinto"]
description = "Django + PostgreSQL + Docker"
license = "MIT"
name = "to_do_or_not_to_do"
version = "2024.09.19"


[tool.poetry.dependencies]
devtools = {extras = ["pygments"], version = "0.11.0"}
dj-database-url = "2.0.0"
django = "4.2.3"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "21.2.0"
psycopg2-binary = "2.9.6"
pydantic = "2.1.0"
pydantic-settings = "2.0.2"
python = "3.11.7"
requests = "2.31.0"
whitenoise = {extras = ["brotli"], version = "6.5.0"}

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.0"
black = "^23.7.0"
django-stubs = "^4.2.3"
djangorestframework-stubs = "^3.14.2"
factory-boy = "^3.3.1"
flake8 = "^6.0.0"
flake8-annotations-complexity = "^0.0.8"
flake8-bandit = "^4.1.1"
flake8-black = "^0.3.6"
flake8-bugbear = "^23.7.10"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.14.0"
flake8-debugger = "^4.1.2"
flake8-executable = "^2.1.3"
flake8-expression-complexity = "^0.0.11"
flake8-fixme = "^1.1.1"
flake8-isort = "^6.0.0"
flake8-multiline-containers = "^0.0.19"
flake8-mutable = "^1.2.0"
flake8-pep3101 = "^2.0.0"
flake8-print = "^5.0.0"
flake8-printf-formatting = "^1.1.2"
flake8-simplify = "^0.20.0"
flake8-variables-names = "^0.0.6"
freezegun = "^1.5.1"
httpie = "^3.2.2"
ipython = "^8.14.0"
isort = "^5.12.0"
mypy = "^1.4.1"
pep8-naming = "^0.13.3"
pytest = "^7.4.0"
types-psycopg2 = "^2.9.21.11"


[tool.black]
line-length = 79
target-version = ["py311"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.local.*
    | .*\.idea.*
    | .*\.vscode.*
)
'''


[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".local",
]


[tool.pytest.ini_options]
cache_dir = ".local/pytest/"
console_output_style = "count"
filterwarnings = [
    "ignore::DeprecationWarning",
]
testpaths = [
    "tests",
]


[tool.mypy]
cache_dir = ".local/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
exclude=[".local/*",]
files = ["."]
no_implicit_optional = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
    "pydantic.mypy",
]
pretty = true
python_version = "3.11"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "pyodbc.*",
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.django-stubs]
django_settings_module = "project.settings"


[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
